ui: Fix bug regarding the displayed state of position
Post positions have displayed states:
- Posted
- Open
- Closed

These states depend on the dates "ends_at", "starts_at" and their comparison
with the current date (now).

If the position is a past instant of a position, these dates should be compared
with the date "updated_at" (instead of now), which is the date that a new
instance of a posted position has been created.