Fix bug with related_positions for user with no permission
When a user has permission to view a position but does not have
permission to view the related_positions (i.e. when a candidate has
applied for one position and the submission date has passed) the
related_positions field in position details fieldset would crash.
This has been resolved by returning in any case the the position model
even if the call for the related positions has failed.