ui: Prevent use of window.open in async context
some browsers prevent use of window.open outside the scope of a click
event. This applies in async url resolving of apella files. User click
triggers the url resolver, once resolved we call window.open, the action
takes place outside the initial click event context and thus browser
ignores the call. This commit mitigates this issue by opening a new
blank window within the click event scope. We keep reference to the
window object and once the url is resolved we use winRef.location to
the file url. This is tested in latest versions of Chrome/Safari. The
introduced process may degrade UX regarding the temporary blank window
appearance. The UX impact is lessened as long as download urls are
instantly resolved.