ui: Fix bug with invisible candidacy fieldset in position
A user would not be able to view candidacies fieldset in position
details view if:
- The user had role "candidate"
- The user had already submitted at least one cancelled candidacy for
  the position
- The user had previously visitied positions-recent view

The bug was caused by misunderstood use of lodash sortBy() function.
We wanted to an array of candidacy objects by updated_at date but
lodash sortBy() uses alpharethmetic sorting so the sorted result
was erroneous.
Custom arithmetic sorting fixed the problem.